Move types from custom named files to types.ts for each module
