+++ pkg/ts/recipe/StoreCook.ts
@@ -9,9 +9,9 @@
 import {logFactory} from '../utils/log.js';
 import {matches} from '../utils/matching.js';
 import {Runtime} from '../Runtime.js';
-import {StoreSpec, Plan} from './Specs.js';
+import {StoreSpec, Plan} from './types.js';
 import {Arc} from '../core/Arc.js';
-import {StoreMeta} from '../core/Meta.js';
+import {StoreMeta} from '../core/types.js';
 
 const log = logFactory(logFactory.flags.recipe, 'StoreCook', '#187e13');
